Move long tests to the intergationtest package
Long tests for Subunit clustering, quaternary symmetry detection and
quaternary structure alignment with some real structure examples have
been moved to the integrationtest module, so that they can be easily
skipped with a configuration option.
